urlShort/requirements.txt
apturl==0.5.2
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.6
django-appconf==1.0.5
django-cryptography==1.1
djangoencryptfile @ https://github.com/ruddra/django-encrypt-file/tarball/0.7
duplicity==0.8.12.0
entrypoints==0.3
executing==1.2.0
fasteners==0.14.1
future==0.18.2
httplib2==0.14.0
idna==2.8
ipdb==0.13.11
ipython==8.10.0
jedi==0.18.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib-inline==0.1.6
monotonic==1.5
mysql-connector-python==8.0.32
mysqlclient==2.1.1
netifaces==0.10.4
numpy==1.24.2
oauthlib==3.1.0
pandas==1.5.3
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
pipdeptree==2.3.3
prompt-toolkit==3.0.36
protobuf==3.20.3
pure-eval==0.2.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.14.0
PyGObject==3.36.0
PyJWT==1.7.1
PyNaCl==1.3.0
python-apt==2.0.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.7.1
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.3
stack-data==0.6.2
systemd-python==234
tomli==2.0.1
traitlets==5.9.0
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.6
xkit==0.0.0

djangoencryptfile~=0.7